# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Shop Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryList"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Products ────────────────────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get customer order history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /customers/{id}/transactions:
    get:
      operationId: getCustomerTransactions
      tags: [customers]
      summary: Get customer transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Customer transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /customers/{id}/disable:
    post:
      operationId: disableCustomer
      tags: [customers]
      summary: Disable a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{id}/add-balance:
    post:
      operationId: addCustomerBalance
      tags: [customers]
      summary: Add balance to customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBalanceRequest"
      responses:
        "200":
          description: Balance added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Promotions ───────────────────────────────────────────────────────────────
...
  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get bot configuration
      responses:
        "200":
          description: Bot config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    post:
      operationId: saveBotConfig
      tags: [bot]
      summary: Save bot configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveBotConfigRequest"
      responses:
        "200":
          description: Config saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
  /bot/test-token:
    post:
      operationId: testBotToken
      tags: [bot]
      summary: Test bot token with Telegram API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestTokenRequest"
      responses:
        "200":
          description: Token test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestTokenResponse"
...
  /bot/webhook:
    post:
      operationId: handleBotWebhook
      tags: [bot]
      summary: Telegram webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Bot Logs ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
    CategoryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Category"
      required:
        - data
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
      required:
        - name
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderCode:
          type: string
        customerId:
          type: integer
        totalAmount:
          type: string
        status:
          type: string
        paymentReference:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderCode
        - customerId
        - totalAmount
        - status
        - createdAt
        - updatedAt
...
    OrderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Transaction:
      type: object
      properties:
        id:
          type: integer
        transactionCode:
          type: string
        paymentReference:
          type: string
          nullable: true
        type:
          type: string
        orderId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        amount:
          type: string
        status:
          type: string
        provider:
          type: string
          nullable: true
        rawPayload:
          type: string
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - transactionCode
        - type
        - amount
        - status
        - createdAt
        - updatedAt
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Customer:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        balance:
          type: string
        totalSpent:
          type: string
        totalOrders:
          type: integer
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - chatId
        - balance
        - totalSpent
        - totalOrders
        - isActive
        - createdAt
        - updatedAt
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    AddBalanceRequest:
      type: object
      properties:
        amount:
          type: string
        note:
          type: string
      required:
        - amount
...
    BotConfig:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        botToken:
          type: string
          nullable: true
        botUsername:
          type: string
          nullable: true
        webhookUrl:
          type: string
          nullable: true
        isConnected:
          type: boolean
        webhookStatus:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - isConnected
    SaveBotConfigRequest:
      type: object
      properties:
        botToken:
          type: string
      required:
        - botToken
    TestTokenRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    TestTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
        username:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - valid
...